requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
Deprecated==1.2.13
Django==4.1.4
django-rest-knox==4.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
greenlet==2.0.1
idna==3.4
kombu==5.2.4
numpy==1.24.1
packaging==22.0
pandas==1.5.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.4.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
toolkit4life==0.1.30
trino==0.320.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1